apps/scraper/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile 'apps\scraper\requirements.in'
#
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   httpx
    #   starlette
beautifulsoup4==4.14.2
    # via -r apps/scraper/requirements.in
certifi==2025.10.5
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.3
    # via requests
click==8.3.0
    # via uvicorn
colorama==0.4.6
    # via
    #   click
    #   tqdm
cryptography==46.0.2
    # via pyjwt
deprecation==2.1.0
    # via
    #   postgrest
    #   storage3
fastapi==0.119.0
    # via -r apps/scraper/requirements.in
filelock==3.20.0
    # via
    #   huggingface-hub
    #   torch
    #   transformers
fsspec==2025.9.0
    # via
    #   huggingface-hub
    #   torch
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
h2==4.3.0
    # via httpx
hpack==4.1.0
    # via h2
httpcore==1.0.9
    # via httpx
httpx[http2]==0.28.1
    # via
    #   postgrest
    #   storage3
    #   supabase
    #   supabase-auth
    #   supabase-functions
huggingface-hub==0.35.3
    # via
    #   tokenizers
    #   transformers
hyperframe==6.1.0
    # via h2
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jinja2==3.1.6
    # via torch
markupsafe==3.0.3
    # via jinja2
mpmath==1.3.0
    # via sympy
multidict==6.7.0
    # via yarl
networkx==3.5
    # via torch
numpy==2.3.3
    # via transformers
packaging==25.0
    # via
    #   deprecation
    #   huggingface-hub
    #   transformers
postgrest==2.22.0
    # via supabase
propcache==0.4.1
    # via yarl
pycparser==2.23
    # via cffi
pydantic==2.12.0
    # via
    #   fastapi
    #   postgrest
    #   pydantic-settings
    #   realtime
    #   storage3
    #   supabase-auth
pydantic-core==2.41.1
    # via pydantic
pydantic-settings==2.11.0
    # via -r apps/scraper/requirements.in
pyjwt[crypto]==2.10.1
    # via supabase-auth
python-dotenv==1.1.1
    # via
    #   -r apps/scraper/requirements.in
    #   pydantic-settings
pyyaml==6.0.3
    # via
    #   huggingface-hub
    #   transformers
realtime==2.22.0
    # via supabase
regex==2025.9.18
    # via transformers
requests==2.32.5
    # via
    #   -r apps/scraper/requirements.in
    #   huggingface-hub
    #   transformers
safetensors==0.6.2
    # via transformers
sniffio==1.3.1
    # via anyio
soupsieve==2.8
    # via beautifulsoup4
starlette==0.48.0
    # via fastapi
storage3==2.22.0
    # via supabase
strenum==0.4.15
    # via supabase-functions
supabase==2.22.0
    # via -r apps/scraper/requirements.in
supabase-auth==2.22.0
    # via supabase
supabase-functions==2.22.0
    # via supabase
sympy==1.14.0
    # via torch
tokenizers==0.22.1
    # via transformers
torch==2.8.0
    # via -r apps/scraper/requirements.in
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   transformers
transformers==4.57.0
    # via -r apps/scraper/requirements.in
typing-extensions==4.15.0
    # via
    #   beautifulsoup4
    #   fastapi
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   realtime
    #   torch
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.5.0
    # via requests
uvicorn==0.37.0
    # via -r apps/scraper/requirements.in
websockets==15.0.1
    # via realtime
yarl==1.22.0
    # via
    #   postgrest
    #   storage3
    #   supabase-functions

# The following packages are considered to be unsafe in a requirements file:
# setuptools
